GFS2: directly write blocks past i_size
Patchwork-id: 34431
O-Subject: [RHEL 6.1 PATCH v2] GFS2: directly write blocks past i_size
Bugzilla: 684115
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Phillip Lougher <plougher@redhat.com>

GFS2 was relying on the writepage code to write out the zeroed data for
fallocate.  However, with FALLOC_FL_KEEP_SIZE set, this may be past i_size.
If it is, it will be ignored.  To work around this, gfs2 now calls
write_dirty_buffer directly on the buffer_heads when FALLOC_FL_KEEP_SIZE
is set, and it's writing past i_size.

This patch is just a cleanup of my last version

This fix has been submitted to Steve's upstream git tree
This fixes bz #684115

This patch applies on top of:
[RHEL 6.1 PATCH] GFS2: fix block allocation check for fallocate (BZ#674603)

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>