UV: Correct kABI from upstream
Patchwork-id: 34661
O-Subject: [PATCH RHEL6.1 BZ684957] UV: Correct kABI from upstream
Bugzilla: 684957
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

BZ#684957 RHEL6.1-Alpha: kABI breakage on UV
First Post

Fix KABI breakage caused by adding a new field to uv_hub_info_s.
The fix is to put the new field in a new structure.

NOTE: The upstream patch that cuased the problem is
    c8f730b1ab825f06733e1c074264f0078721f365
The modules for which this is a problem are primarily SGI's.

All changes are to UV specific code.

Signed-off-by: Russ Anderson <rja@sgi.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>