perf record: Add option to disable collecting build-ids
Patchwork-id: 34847
O-Subject: [PATCH RHEL6.1 005/120] perf record: Add option to disable collecting
	build-ids
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
baa2f6c perf record: Add option to disable collecting build-ids

Collecting build-ids for long running sessions may take a long time
because it needs to traverse the whole just collected perf.data stream
of events, marking the DSOs that had hits and then looking for the
.note.gnu.build-id ELF section.

For things like the 'trace' tool that records and right away consumes
the data on systems where its unlikely that the DSOs being monitored
will change while 'trace' runs, it is desirable to remove build id
collection, so add a -B/--no-buildid option to perf record to allow such
use case.

Longer term we'll avoid all this if we, at DSO load time, in the kernel,
take advantage of this slow code path to collect the build-id and stash
it somewhere, so that we can insert it in the PERF_RECORD_MMAP event.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>