perf tools: Fix lost and unknown events handling
Patchwork-id: 34853
O-Subject: [PATCH RHEL6.1 009/120] perf tools: Fix lost and unknown events
	handling
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
068ffaa perf tools: Fix lost and unknown events handling

Fix it by explaining what can be happening and giving the number of processed
and lost events.

Also holler if unknown events were found, that can be due to processing a
perf.data file collected using a newer tool where newer events got added on
reporting using an older perf tool, that or a bug, so ask for a report to be
made.

Works on both --tui and --stdio.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>