perf session: Fix list sort algorithm
Patchwork-id: 34854
O-Subject: [PATCH RHEL6.1 010/120] perf session: Fix list sort algorithm
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
a1225de perf session: Fix list sort algorithm

The homebrewn sort algorithm fails to sort in time order. One of the problem
spots is that it fails to deal with equal timestamps correctly.

My first gut reaction was to replace the fancy list with an rbtree, but the
performance is 3 times worse.

Rewrite it so it works.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>