perf session: Keep file mmaped instead of malloc/memcpy
Patchwork-id: 34857
O-Subject: [PATCH RHEL6.1 016/120] perf session: Keep file mmaped instead of
	malloc/memcpy
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
fe17420 perf session: Keep file mmaped instead of malloc/memcpy

Profiling perf with perf revealed that a large part of the processing time is
spent in malloc/memcpy/free in the sample ordering code. That code copies the
data from the mmap into malloc'ed memory. That's silly. We can keep the mmap
and just store the pointer in the queuing data structure. For 64 bit this is
not a problem as we map the whole file anyway. On 32bit we keep 8 maps around
and unmap the oldest before mmaping the next chunk of the file.

Performance gain: 2.95s -> 1.23s (Faktor 2.4)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>