perf script: Fix compiler warning in builtin_script.c:is_top_script()
Patchwork-id: 34901
O-Subject: [PATCH RHEL6.1 042/120] perf script: Fix compiler warning in
	builtin_script.c:is_top_script()
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
965bb6b perf script: Fix compiler warning in builtin_script.c:is_top_script()

Fix annoying compiler warning in the is_top_script() function.

The issue was that a const char * was cast into a char * to call
ends_with(). We fix the users of ends_with() instead. Some are passing a
char *, but it is okay to cast the return value of ends_with() to char *
(because we understand what ends_with() does).

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>