perf session: Fallback to unordered processing if no sample_id_all
Patchwork-id: 34913
O-Subject: [PATCH RHEL6.1 056/120] perf session: Fallback to unordered
	processing if no sample_id_all
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
21ef97f perf session: Fallback to unordered processing if no sample_id_all

If we are running the new perf on an old kernel without support for
sample_id_all, we should fall back to the old unordered processing of
events. If we didn't than we would *always* process events without
timestamps out of order, whether or not we hit a reordering race. In
other words, instead of there being a chance of not attributing samples
correctly, we would guarantee that samples would not be attributed.

While processing all events without timestamps before events with
timestamps may seem like an intuitive solution, it falls down as
PERF_RECORD_EXIT events would also be processed before any samples.
Even with a workaround for that case, samples before/after an exec would
not be attributed correctly.

This patch allows commands to indicate whether they need to fall back to
unordered processing, so that commands that do not care about timestamps
on every event will not be affected. If we do fallback, this will print
out a warning if report -D was invoked.

This patch adds the test in perf_session__new so that we only need to
test once per session. Commands that do not use an event_ops (such as
record and top) can simply pass NULL in it's place.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>