perf probe: Don't always consider EOF as an error when listing source code
Patchwork-id: 34898
O-Subject: [PATCH RHEL6.1 064/120] perf probe: Don't always consider EOF as an
	error when listing source code
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
fde52db perf probe: Don't always consider EOF as an error when listing source code

When listing a whole file or a function which is located at the end,
perf-probe -L output wrongly: "Source file is shorter than expected.".

This is because show_one_line() always consider EOF as an error.

This patch fixes this by not considering EOF as an error when dumping
the trailing lines. Otherwise it's still an error and perf-probe still
outputs its warning.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>