perf top: Fix annotate segv
Patchwork-id: 34942
O-Subject: [PATCH RHEL6.1 100/120] perf top: Fix annotate segv
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
cc84158 perf top: Fix annotate segv

Before we had sym_counter, it was initialized to zero and we used that
as an index in the global attrs variable, now we have a list of evsel
entries, and sym_counter became sym_evsel, that remained initialized to
zero (NULL): b00m.

Fix it by initializing it to the first entry in the evsel list.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>