perf tools: Fix build by checking if extra warnings are supported
Patchwork-id: 34958
O-Subject: [PATCH RHEL6.1 112/120] perf tools: Fix build by checking if extra
	warnings are supported
Bugzilla: 664900

Bugzilla: 664900
https://bugzilla.redhat.com/show_bug.cgi?id=664900

backport of:
065bef5 perf tools: Fix build by checking if extra warnings are supported

The -Wstack-protector and -Wvolatile-register-var warnings, for
instance, are not supported by gcc 3.4.6.

So fix by doing the same check we already do for -fstack-protector-all.

With this and the other patches in this series, perf builds unmodified
on, for instance, RHEL4.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>