nmi_watchdog: compile-in i686 but disable by default
Patchwork-id: 34988
O-Subject: [RHEL6 PATCH 1/2] i686,
	nmi_watchdog: compile-in but disable by default
Bugzilla: 685021
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=685021

When hooking in the new nmi_watchdog into the config options, I disabled
compiling it on i686 because we didn't support it on that platform by
default.  However, I forgot that people can still manually enable it using
nmi_watchdog=1/2 on the command line.

Currently when people do that it uses the old nmi_watchdog which conflicts
with the perf subsystem and has other issues.

This patch correctly complies in the new nmi_watchdog for i686 platforms but
adds the option to leave it off by default for that arch.  Normally the
nmi_watchdog is just enabled by default if compiled in.  This patch changes
that behaviour and therefore I have to re-add all the expected nmi_watchdog
enable options like 'nmi_watchdog=1/2/lapic/ioapic'.

I won't be pushing this patch upstream as technically it is kinda of an odd
hack (enable support but keep it off by default).  So this will be a RHEL-6
only patch for now.  Upstream has i686 enabled by default.

Tested locally on an i686 and x86_64 machine.

Signed-off-by: Don Zickus <dzickus@redhat.com>