spec: add phy drivers to initrd
Patchwork-id: 34998
O-Subject: [PATCH RHEL6.1] specfile: add phy drivers to initrd
Bugzilla: 650907
RH-Acked-by: John Feeney <jfeeney@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Users reported that Broadcom 57780 NICs (tg3-based) were not able to PXE
boot and install RHEL6 over those devices.  This was because the phy
drivers (in this case broadcom.ko) were not included in the initrd.  I
noticed that the kernel specfile for Fedora also keys off the
'phy_driver_register' symbol to know that we should include a module as
part of networking, and that seems like a reasonable solution to me.

Tested by me to resolve RHBZ 650907.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>