ptrace: Remove BUG_ON when full register set not available
Patchwork-id: 35044
O-Subject: [PATCH RHEL6.1 BZ678099] powerpc/ptrace: Remove BUG_ON when full
	register set not available
Bugzilla: 678099
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=678099

Description:
------------
In some cases during a threaded core dump not all the threads will have
a full register set. This happens when the signal causing the core dump
races with a thread exiting.  The race happens when the exiting thread
has entered the kernel for the last time before the signal arrives, but
doesn't get far enough through the exit code to avoid being included
in the core dump.

So we get a thread included in the core dump which is never going to go
out to userspace again and only has a partial register set recorded

Normally we would catch each thread as it is about to go into userspace
and capture the full register set then.

However, this exiting thread is never going to go out to userspace
again, so we have no way to capture its full register set.  It doesn't
really matter, though, as this is a thread which is effectively
already dead.

So instead of hitting a BUG() in this case (a really bad choice of
action in the first place), we use a poison value for the register
values.

[BenH]: Some cosmetic/stylistic changes and fix build on ppc32

Signed-off-by: Mike Wolf <mjw@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
RHEL 6.0

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3202791

Upstream:
---------
Commit: http://git.kernel.org/?p=linux/kernel/git/benh/powerpc.git;
a=commitdiff;h=a71f5d5d279375205009a4be56a3cf6682921292

Test Status:
------------

Tested by the IBM test team.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>