Fix regression with SMP guests
Patchwork-id: 35045
O-Subject: [PATCH RHEL 6.1 URGENT] Fix regression with SMP guests
Bugzilla: 681133
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Glauber Costa <glommer@redhat.com>

I'm actually rather unhappy with this patch.  If fixes a regression, but
it took me a long time to arrive at good understanding why it does so
and I'm still not completely satisfied with the explanation.  It means
there are further subtle problems with KVM clock that we are not yet
adequately addressing in RHEL 6 (or even upstream).

Nevertheless, the bug is easy to verify, this patch is from upstream
code, it does make things work, and it restores the old conservative
behavior of requesting a KVM clock update every time a VCPU switches or
loads on a different physical CPU.

The problem with this patch is that whenever a KVM clock is registered,
we will already request a KVM clock update - KVM clock data is a per-cpu
item thus already done by necessity on each CPU.  So what is this patch
really fixing?

The only other reason to request a KVM clock update is if we have
unsynchronized TSCs - and this case was handled already prior to this
change.  So there appears to be no case in which this patch will
actually help to fix any known bug - unless we are missing a case where
TSC has become desynchronized or something else has gone wrong.  A
physical CPU switch on a system with a stable TSC should not require any
further correction to KVM clock.  With the new KVM clock patches in RHEL
6, on synchronized systems, there is no longer a TSC offset difference
across CPUs, however there is still a small but perceptible delta
between KVM clock on multiple VCPUs.  This happens because the
measurement of system time for each per-cpu kvmclock area is not
completely simultaneous with the measurment of the TSC.  This may cause
slight forward and backwards jumps of KVM clock time when read globally
across different CPUs, which could conceivably cause a regression like this.

However this patch doesn't actually fix that problem.  It simply
restores the old behavior of updating KVM clock data more frequently -
at least at every PCPU switch - which could simply be papering over the
actual root cause of this bug.

I have two viable theories about what the root cause might be and hope
to get confirmation upstream.  For now, this patch is the best that can
be done, and should at least stop this bug from remaining a test-stopper.

Please ack and apply.

Zach
>From 0ec7b09af2ed1480a788dd3c67d38b00b039813d Mon Sep 17 00:00:00 2001
From: Zachary Amsden <zamsden@redhat.com>
Date: Thu, 24 Mar 2011 17:52:07 -0400
Subject: [RHEL6.1] Fix KVM clock regression

BZ: 681113

Backport of upstream commit 1aa8ceef0312a6aae7dd863a120a55f1637b361d
Author: Nikola Ciprich <extmaillist@linuxbox.cz>
Date:   Wed Mar 9 23:36:51 2011 +0100

    KVM: fix kvmclock regression due to missing clock update

    commit 387b9f97750444728962b236987fbe8ee8cc4f8c moved kvm_request_guest_time
    breaking 32bit SMP guests using kvm-clock. Fix this by moving (new) clock up
    to proper place.

Signed-off-by: Zachary Amsden <zamsden@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>