svcrpc: take lock on turning entry NEGATIVE in cache_check
Patchwork-id: 35083
O-Subject: [RHEL6.1 PATCH 3/5] svcrpc: take lock on turning entry NEGATIVE in
	cache_check
Bugzilla: 690900
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

We attempt to turn a cache entry negative in place.  But that entry may
already have been filled in by some other task since we last checked
whether it was valid, so we could be modifying an already-valid entry.
If nothing else there's a likely leak in such a case when the entry is
eventually put() and contents are not freed because it has
CACHE_NEGATIVE set.

So, take the cache_lock just as sunrpc_cache_update() does.

Reviewed-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>

Bugzilla 690900
Upstream 6bab93f87ec703bf6650875881b11f9f27d7da56

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>