NFS: Fix a hang/infinite loop in nfs_wb_page()
Patchwork-id: 35117
O-Subject: [RHEL6.1 PATCH 1/1] NFS: Fix a hang/infinite loop in nfs_wb_page()
Bugzilla: 672305
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

When one of the two waits in nfs_commit_inode() is interrupted, it
returns a non-negative value, which causes nfs_wb_page() to think
that the operation was successful causing it to busy-loop rather
than exiting.
It also causes nfs_file_fsync() to incorrectly report the file as
being successfully committed to disk.

This patch fixes both problems by ensuring that we return an error
if the attempts to wait fail.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Cc: stable@kernel.org
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>