ipsec: Disable granular bundles
Patchwork-id: 35119
O-Subject: [RHEL6 PATCH] ipsec: Disable granular bundles
Bugzilla: 631833
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

RHEL6 Bugzilla #631833

ipsec: Disable granular bundles

The feature XFRM_SUB_POLICY causes all bundles to be at the finest
granularity possible, which unfortunately the RHEL6 kernel cannot
handle correctly.  Because the data structure used in RHEL6 for this
is a linked list, we end up with at best a serious performance
problem and at worst a DoS opportunity.

Upstream has reworked the flow cache to handle this properly.

At this point in time the risk/cost in backporting that is too
great compared to the problem we have.

So this patch simply disables the part of XFRM_SUB_POLICY that
creates the problematic bundles by simply reverting to previous
behaviour.

This obviously may affect the functionality of XFRM_SUB_POLICY
but only in the cases where sub-IP address granularity is needed.

Tested-by: Adam Okuliar <aokuliar@redhat.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

index cb81ca3..b526875 100644
Cheers,

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>