bnx2fc: common free list for cleanup commands
Patchwork-id: 35140
O-Subject: [RHEL6.1 PATCH 04/14] bnx2fc: common free list for cleanup commands
Bugzilla: 683153
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Bhanu Gollapudi <bprakash@broadcom.com>

This is in the scsi-misc tree commit 0ea5c27583e1cc164bba7ca29fe48a225f52d19b

Cleanup commands are issued to the firmware to cleanup any stuck ios
that are supposed to be implicitly aborted. In the worst case we can
have all scsi ios filling up the free_list and we may not be able to
allocate cleanup tasks. So the driver has to reserve free_list entries
to be able to allocate the cleanup tasks. This reserve free_list common
to all cpus is allocated as one additional entry in the per cpu
free_lists.

In bnx2fc_cmd_alloc(), there is a related fix to use get_cpu() for the
free_list_index.  This will prevent using the wrong index if the CPU
is preempted.

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Michael Chan <mchan@broadcom.com>
Signed-off-by: James Bottomley <James.Bottomley@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>