bnx2fc: Fix kernel panic when deleting NPIV ports
Patchwork-id: 35152
O-Subject: [RHEL6.1 PATCH 13/14] bnx2fc: Fix kernel panic when deleting NPIV
	ports
Bugzilla: 683153
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Bhanu Prakash Gollapudi <bprakash@broadcom.com>

Deleting NPIV port causes a kernel panic when the NPIV port is in the same zone
as the physical port and shares the same LUN. This happens due to the fact that
vport destroy and unsolicited ELS are scheduled to run on the same workqueue,
and vport destroy destroys the lport and the unsolicited ELS tries to access
the invalid lport.  This patch fixes this issue by maintaining a list of valid
lports and verifying if the lport is valid or not before accessing it.

This is not upstream. It has been sent
http://marc.info/?l=linux-scsi&m=130135399825872&w=2

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>