Be more careful about clearing flags bit in ->recovery
Patchwork-id: 35249
O-Subject: [PATCH 11/37] md: Be more careful about clearing flags bit in
	->recovery
Bugzilla: 688725
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

From: NeilBrown <neilb@suse.de>

Setting ->recovery to 0 is generally not a good idea as it could clear
bits that shouldn't be cleared.  In particular, MD_RECOVERY_FROZEN
should only be cleared on explicit request from user-space.

So when we need to clear things, just clear the bits that need
clearing.

As there are a few different places which reap a resync process - and
some do an incomplte job - factor out the code for doing the from
md_check_recovery and call that function instead of open coding part
of it.

upstream: 7ebc0be7fff4146e87b4078f054977b72998abd3

Signed-off-by: NeilBrown <neilb@suse.de>
Reported-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>