Fix removal of extra drives when converting RAID6 to RAID5
Patchwork-id: 35270
O-Subject: [PATCH 20/37] md: Fix removal of extra drives when converting RAID6
	to RAID5
Bugzilla: 688725
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

From: NeilBrown <neilb@suse.de>

When a RAID6 is converted to a RAID5, the extra drive should
be discarded.  However it isn't due to a typo in a comparison.

This bug was introduced in commit e93f68a1fc6 in 2.6.35-rc4
and is suitable for any -stable since than.

As the extra drive is not removed, the 'degraded' counter is wrong and
so the RAID5 will not respond correctly to a subsequent failure.

upstream: bf2cb0dab8c97f00a71875d9b13dbac17a2f47ca

Cc: stable@kernel.org
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>