perf: let everyone share counters on a P4 machine
Patchwork-id: 35232
O-Subject: [RHEL6 PATCH 1/2] perf,
	x86: let everyone share counters on a P4 machine
Bugzilla: 688547
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=688547

Currently, the perf nmi handler on Pentium4 machines only let one user
access it registers, either the perf tool or the nmi watchdog.  This is
silly because the nmi watchdog is a consumer of the perf subsystem, they
should be able to share.

Because of hw register restrictions, they don't.  This patch changes that so
you can have the nmi watchdog running in the back ground and run 'perf' in
the foreground.  To accomplish this, the registers the nmi watchdog uses to
do its calculations is changed from non-halted ticks, to non-sleeping ticks.
As a result, this use case is slightly different than what the perf
subsystem initially thought.  Therefore a new hardware event was used.

This has been posted privately and tested internally by myself.  It isn't
the ideal patch to fix this situation but given the deadlines and
regression, it should be good enough for now.  Hopefully, I can have a
better fix for 6.2.  Tracking that with bz692677.

It hasn't been posted upstream because it causes more unknown NMIs as a
result of more NMI activity.  The next patch will temporarily resolve that
too until RHEL-6.2.

	From: Cyrill Gorcunov <gorcunov@gmail.com>
	Date: Sun, 27 Mar 2011 19:58:30 +0400
	Subject: [PATCH 2/2] perf, x86: Add PERF_COUNT_HW_NMI_WATCHDOG event

	Due to restriction and specifics of P4 PMU we need
	a separated event for NMI watchdog. Note that on
	all other than P4 PMU this event is an alias for
	PERF_COUNT_HW_CPU_CYCLES.

	Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>