fix the test for finding spares in raid5_start_reshape
Patchwork-id: 35261
O-Subject: [PATCH 23/37] md: fix the test for finding spares in
	raid5_start_reshape.
Bugzilla: 688725
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

From: NeilBrown <neilb@suse.de>

As spares can be added to the array before the reshape is started,
we need to find and count them when checking there are enough.
The array could have been degraded, so we need to check all devices,
no just those out side of the range of devices in the array before
the reshape.

So instead of checking the index, check the In_sync flag as that
reliably tells if the device is a spare or this purpose.

upstream: 469518a3455c79619e9231aeffeffa2e2989f738

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>