quota: Don't write quota info in dquot_commit()
Patchwork-id: 35308
O-Subject: [RHEL6.1 PATCH 2/2 v2] quota: Don't write quota info in dquot_commit()
Bugzilla: 680105
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Jan Kara <jack@suse.cz>

BZ 680105
https://bugzilla.redhat.com/show_bug.cgi?id=680105

Brew 3221278
https://brewweb.devel.redhat.com/taskinfo?taskID=3221278

This commit is not upstream yet, but it has been posted and is waiting
for inclusion.

Tested with xfstest and stress tested with xfstest 234 in the loop with
sync in the other loop. Works as expected

There's no reason to write quota info in dquot_commit(). The writing is a
relict from the old days when we didn't have dquot_acquire() and
dquot_release() and thus dquot_commit() could have created / removed quota
structures from the file. These days dquot_commit() only updates usage counters
/ limits in quota structure and thus there's no need to write quota info.

This also fixes an issue with journaling filesystem which didn't reserve
enough space in the transaction for write of quota info (it could have been
dirty at the time of dquot_commit() because of a race with other operation
changing it).

Reported-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Jan Kara <jack@suse.cz>